requirements.txt
hydra-core~=1.2.0
nmslib~=2.1.1
numpy~=1.22.0
omegaconf~=2.2.2
pandas~=1.4.3
pytorch-lightning==1.6.5
ranx~=0.3.6
torch==1.13.1
torchmetrics~=0.9.2
tqdm~=4.64.0
transformers~=4.27.2
